requirements.txt
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
structlog>=23.2.0
jira>=3.5.0
pydantic>=2.4.2
prometheus-client>=0.17.1
httpx>=0.25.0
pytest>=7.4.3 